Moving a couple classes to android package.
