+++ formula-android/src/main/java/com/instacart/formula/fragment/FormulaFragment.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.instacart.formula.Cancelable
 import com.instacart.formula.android.FeatureView
+import com.instacart.formula.android.FragmentEnvironment
 import com.instacart.formula.android.FragmentKey
 import com.instacart.formula.android.ViewFactory
 import com.jakewharton.rxrelay3.BehaviorRelay
