+++ formula-android/src/test/java/com/instacart/formula/android/ActivityStoreFactoryTest.kt
@@ -3,7 +3,6 @@ package com.instacart.formula.android
 import androidx.fragment.app.FragmentActivity
 import com.google.common.truth.Truth.assertThat
 import com.instacart.formula.android.internal.ActivityStoreFactory
-import com.instacart.formula.fragment.FragmentEnvironment
 import com.nhaarman.mockito_kotlin.mock
 import org.junit.Test
