+++ formula-android/src/test/java/com/instacart/formula/android/FragmentFlowStoreTest.kt
@@ -6,7 +6,6 @@ import com.instacart.formula.android.fakes.FakeAuthFlowFactory
 import com.instacart.formula.android.fakes.FakeComponent
 import com.instacart.formula.android.fakes.MainKey
 import com.instacart.formula.fragment.FragmentContract
-import com.instacart.formula.fragment.FragmentEnvironment
 import com.instacart.formula.fragment.FragmentFlowStore
 import com.instacart.formula.fragment.FragmentLifecycleEvent
 import com.instacart.formula.android.fakes.TestAccountFragmentContract
